+++ src/main/java/de/einholz/ehdynview/config/DefConfig.java
@@ -11,7 +11,7 @@ public final class DefConfig {
     public static final Text FPS_MIN_COMMENT = Text.translatable("config.ehdynview.fps_min.comment");
     public static final int BUFFER_SIZE_VALUE = 10;
     public static final Text BUFFER_SIZE_COMMENT = Text.translatable("config.ehdynview.buffer_size.comment");
-    public static final int SAMPLING_PERIOD_VALUE = 2000000000; // in nanoseconds here: 2s
+    public static final long SAMPLING_PERIOD_VALUE = 2000000000; // in nanoseconds here: 2s
     public static final Text SAMPLING_PERIOD_COMMENT = Text.translatable("config.ehdynview.sampling_period.comment");
 
     private DefConfig() {}
@@ -21,7 +21,7 @@ public final class DefConfig {
         json.putDefault("fpsMax", new JsonPrimitive((Integer) FPS_MAX_VALUE), null);
         json.putDefault("fpsMin", new JsonPrimitive((Integer) FPS_MIN_VALUE), null);
         json.putDefault("bufferSize", new JsonPrimitive((Integer) BUFFER_SIZE_VALUE), null);
-        json.putDefault("samplingPeriod", new JsonPrimitive((Integer) SAMPLING_PERIOD_VALUE), null);
+        json.putDefault("samplingPeriod", new JsonPrimitive((Long) SAMPLING_PERIOD_VALUE), null);
         return json;
     }
 }
